+++ src/Pim/Component/Connector/Processor/DummyItemProcessor.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\BaseConnectorBundle\Processor;
+namespace Pim\Component\Connector\Processor;
 
 use Akeneo\Bundle\BatchBundle\Item\AbstractConfigurableStepElement;
 use Akeneo\Bundle\BatchBundle\Item\ItemProcessorInterface;
@@ -9,12 +9,10 @@ use Akeneo\Bundle\BatchBundle\Item\ItemProcessorInterface;
  * Dummy step, can be use to do nothing until you'll have concrete implementation
  *
  * @author    Nicolas Dupont <nicolas@akeneo.com>
- * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
+ * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- *
- * @deprecated please use Pim\Component\Connector\Processor\DummyItemProcessor
  */
-class DummyProcessor extends AbstractConfigurableStepElement implements ItemProcessorInterface
+class DummyItemProcessor implements ItemProcessorInterface
 {
     /**
      * {@inheritdoc}
@@ -22,13 +20,5 @@ class DummyProcessor extends AbstractConfigurableStepElement implements ItemProc
     public function process($item)
     {
         return null;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getConfigurationFields()
-    {
-        return array();
     }
 }