+++ src/Pim/Component/Connector/Writer/DummyItemWriter.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\BaseConnectorBundle\Writer;
+namespace Pim\Component\Connector\Writer;
 
 use Akeneo\Bundle\BatchBundle\Item\AbstractConfigurableStepElement;
 use Akeneo\Bundle\BatchBundle\Item\ItemWriterInterface;
@@ -9,12 +9,10 @@ use Akeneo\Bundle\BatchBundle\Item\ItemWriterInterface;
  * Dummy step, can be use to do nothing until you'll have concret implementation
  *
  * @author    Nicolas Dupont <nicolas@akeneo.com>
- * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
+ * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- *
- * @deprecated please use Pim\Component\Connector\Writer\DummyItemWriter
  */
-class DummyWriter extends AbstractConfigurableStepElement implements ItemWriterInterface
+class DummyItemWriter implements ItemWriterInterface
 {
     /**
      * {@inheritdoc}
@@ -22,13 +20,5 @@ class DummyWriter extends AbstractConfigurableStepElement implements ItemWriterI
     public function write(array $items)
     {
         return null;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getConfigurationFields()
-    {
-        return array();
     }
 }